PPU now stores a colour buffer.
Previously the PPU had to read from the frame buffer and reconstruct the
Gameboy colour from the colour in the frame buffer. However if two
colours in the pallet were the same it might give the wrong colour
(which would affect sprites).

Now the PPU stores a colour buffer to not have to recalculate what the
Gameboy colour should be and to have a more accurate colour readout at
the expense of some extra memory usage.